% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc-data-handling.R
\name{generateMNLrandomTab}
\alias{generateMNLrandomTab}
\title{Simple experimental design for choice-based conjoint analysis}
\usage{
generateMNLrandomTab(
  attrLevels,
  cards = 3,
  respondents = 200,
  trials = 12,
  balanced.sample = TRUE,
  best.of = 50,
  verbose = TRUE,
  no.output = FALSE
)
}
\arguments{
\item{attrLevels}{A vector of integers, where each integer is the number of
levels for its respective attribute. For example, if a product has 3
brands, 4 performance levels, and 3 price points, this would be
\code{c(3, 4, 3)}.}

\item{cards}{The number of concepts to show at one time.}

\item{respondents}{The number of designs to generate.}

\item{trials}{How many tasks will be given to each respondent}

\item{balanced.sample}{Whether to attempt to do level balancing across
attributes, i.e., to have fewer duplications of any level on a given
task.}

\item{best.of}{How many iterations to consider before selecting the
optimized one.}

\item{verbose}{Outputs more information as it works.}

\item{no.output}{Whether to suppress all output.}
}
\value{
A data frame for the experimental design, with a row for
each respondent * trial * concept, with columns for the attribute levels.
}
\description{
Create an experimental design for an exactly rectangular version of
choice-based conjoint analysis (same number of concepts and trials
for every respondent).
This function is recommended primarily for simple surveys and
didactic purposes; it does not do advanced optimization.
It will attempt to find level balance through iterative selection across
multiple randomized designs.
}
\seealso{
[generateRNDpws] to create a set of random utilities for
a specific attribute list, [pickMNLwinningCards] to "answer" the survey
design according to utilities.
}
