% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc-data-handling.R
\name{generateRNDpws}
\alias{generateRNDpws}
\title{Generate some zero-centered part worths (utilities) for CBC design}
\usage{
generateRNDpws(attrs, mu = 0, sdev = 1)
}
\arguments{
\item{attrs}{A vector with the attribute/level design, as noted for
\code{generateMNLrandomTab}#'}

\item{mu}{The mean of the random utilities}

\item{sdev}{Standard deviation of the utilities}
}
\value{
A vector of random values that sum to 0 for each subset of values
that correspond to one attribute in a CBC attribute/level design.
}
\description{
Generate some zero-centered part worths (utilities) for CBC design
}
\seealso{
[generateMNLrandomTab] to create a design matrix,
[pickMNLwinningCards] to select the winning concepts from it, given
some set of utilities (as from this function).
}
