% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lepageD.R
\name{lepageD}
\alias{lepageD}
\title{Perform the Lepage Two-Sample Location-Scale Test with Simulation}
\usage{
lepageD(sim, n, modelname1, modelname2, meanname, sdname)
}
\arguments{
\item{sim}{Integer. The number of simulations to run.}

\item{n}{Integer. The number of samples to draw for each simulation.}

\item{modelname1}{A list containing `estimate` (mean vector) and `varcov` (variance-covariance matrix) for the first model. Can be an apollo model object.}

\item{modelname2}{A list containing `estimate` (mean vector) and `varcov` (variance-covariance matrix) for the second model. Can be an apollo model object.}

\item{meanname}{Character. The name of the column representing the mean parameter in the models.}

\item{sdname}{Character. The name of the column representing the standard deviation parameter in the models.}
}
\value{
A list of Lepage test results from each simulation, containing test statistics and p-values.
}
\description{
This function simulates data from two estimated models and applies the 
Lepage two-sample location-scale test to assess differences in location 
and/or scale between two distributions.
}
\details{
The function:
1. Generates parameter draws using `takedraws()`.
2. Simulates normal distributions based on these parameters.
3. Applies the Lepage two-sample location-scale test using `lepage.test()`.
4. Returns a list of results from each simulation.
}
\examples{
model1 <- list(estimate = c(mean = 1.5, sd = 0.5), varcov = matrix(c(0.1, 0.02, 0.02, 0.2), nrow = 2))
model2 <- list(estimate = c(mean = 1.8, sd = 0.6), varcov = matrix(c(0.12, 0.03, 0.03, 0.25), nrow = 2))
results <- lepageD(sim = 100, n = 50, modelname1 = model1, modelname2 = model2, meanname = "mean", sdname = "sd")
head(results)

}
