% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdiff-estimate.R
\name{md.hb}
\alias{md.hb}
\title{Hierarchical Bayes estimation for MaxDiff data}
\usage{
md.hb(
  md.define,
  mcmc.iters = 1000,
  pitersUsed = 0.1,
  mcmc.seed = runif(1, min = 0, max = 1e+08),
  restart = FALSE
)
}
\arguments{
\item{md.define}{The structured data with MaxDiff observations. This is
typically created by an importing function such as \code{read.md.qualtrics()}}

\item{mcmc.iters}{How many iterations to run the MCMC estimation process.
Default is 1000 iterations (suitable only for testing), recommend 10000 or
more for typical usage.}

\item{mcmc.seed}{Random number seed to make the process repeatable. Default
is that the function will draw a random number to be the seed and report it.}

\item{pitersUsers}{The proportion of the MCMC chain to retain, from
the end of the chain. Default 0.1 for 10% posterior draws.}
}
\value{
Returns a list with the following objects: \code{md.model.hb} is the
result from a call to \code{choicemodelr} to estimate the model;
\code{md.hb.betas} are the raw multinomial logit model beta coefficients;
and \code{md.hb.betas.zc} are zero-centered difference scores that may
be more interpretable for stakeholder audiences. Use \code{plot.md.range()}
to plot the aggregate results, or \code{plot.md.indiv()} to plot the
individual-level results, or \code{plot.md.group()} to compare
distributions by categorical groups such as demographic or treatment groups.
}
\description{
Estimates a hierarchical Bayes (HB) model for MaxDiff observations.
This is primarily a wrapper for \code{ChoiceModelR::choicemodelr} that
formats the data, calls \code{choicemodelr}, and extracts the results.
}
