% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdiff-estimate.R
\name{md.quicklogit}
\alias{md.quicklogit}
\title{Estimate MaxDiff utilities quickly with an aggregate logit model}
\usage{
md.quicklogit(md.define, preadapt.only = TRUE)
}
\arguments{
\item{md.define}{A study object as documented for \code{parse.md.qualtrics()}}

\item{preadapt.only}{An experimental parameter for adaptive models, to be
documented in the future. For now, leave this as \code{TRUE}.}
}
\value{
A model object from \code{mlogit::mlogit()} with parameter estimates
for K-1 levels of your MaxDiff items. Use \code{md.plot.logit()} to plot
the results.
}
\description{
\code{md.quicklogit} is intended to give a quick check to see whether
your data are structure correctly and have reasonable estimates,
before investing time to run a hierarchical Bayes model (\code{md.hb()}).
A common error is to have the values reversed for the "best" and "worst"
observations. That will appear in the results with obviously preferred
items showing low preference, while poor items show high preference.

The fix in that case is to relabel those data points such that the
"best" items' value
is greater than the value for the worst items (the exact values don't
matter). For example, if your data accidentally code best as 1, and worst
as 2, you could replace all of the best observations with a value of 3.
Then run \code{md.quicklogit} again.

Note that \code{md.quicklogit} drops 1 item for model identification, and
thus reports K-1 estimates. For example, if you have 15 items you will
see 14 estimated parameters in a summary of the return object.
}
\seealso{
[md.hb()] for the recommended hierarchical Bayes estimation.
}
