% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdiff-data-importexport.R
\name{parse.md.qualtrics}
\alias{parse.md.qualtrics}
\title{Parse Qualtrics MaxDiff data and optionally set up a study object}
\usage{
parse.md.qualtrics(
  file.qsv = NULL,
  itemSplit = "\%-\%",
  designHead = "Display Order",
  itemConfirm = NULL,
  friendly.names = NULL,
  returnList = FALSE
)
}
\arguments{
\item{file.qsv}{The Qualtrics exported QSV (quasi-CSV) file.}

\item{itemSplit}{The short string used by Quatrics to split item headers and text}

\item{designHead}{String used by Qualtrics to identify columns in the design matrix}

\item{itemConfirm}{Optional. A string that appears only in your MaxDiff item
headers (e.g., tje question text) and not in any other items. This can
help the parse routine identify the MaxDiff items.}

\item{friendly.names}{TBW}

\item{returnList}{TBW}
}
\value{
TBW
}
\description{
In a Qualtrics exported data set, at the time of writing, there is no
positive identification of which items in a data set are MaxDiff items vs
other kinds of survey items.
This function attempts to identify MaxDiff items using various tactics,
and also infers properties of the MaxDiff design, such as the number of
items, their text, and the number of task sets.
}
\details{
That information is needed to set up a "study object" that we
typically name \code{md.define}.
This function can optionally set up that object for you, and then the
study object can be used by \code{read.md.qualtrics()} to actually import
the data, as well as by other functions to estimate and plot the
results (e.g., \code{md.hb()} and \code{plot.md.range()}).

Be sure to read all output carefully to make sure it matches your
expectations.
If the import fails, check the above requirements carefully and remove
any respondents with incomplete data. Check that the exported data
set matches the requirements.
Compare the "qualtrics-pizza-maxdiff.csv" data set in this package to
see whether your data appear similar in structure.

Additional details and examples are in Chapter 10 of Chapman & Rodden,
"Quantitative User Experience Research".
}
\seealso{
[read.md.qualtrics] to actually read in a data set,
  [parse.md.cho] for a similar function for Sawtooth Software .CHO files.
}
