% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc-data-handling.R
\name{pickMNLwinningCards}
\alias{pickMNLwinningCards}
\title{Choose the winning concepts in a CBC design matrix, according to utilities}
\usage{
pickMNLwinningCards(
  design.mat,
  pws = rep(0, ncol(design.mat)),
  cards = 3,
  noise = FALSE,
  p.noise = 0.3,
  use.MNL = TRUE,
  verbose = TRUE,
  vec.format = "WIN"
)
}
\arguments{
\item{design.mat}{A design matrix as created by \code{generateMNLrandomTab}}

\item{pws}{A vector of utility coefficients for each column in
\code{design.mat}. If missing, these are set to 0 and all winners will be
drawn randomly.}

\item{cards}{How many concepts are seen on each trial in the design matric.}

\item{noise}{Whether to randomly make some choices (instead of using the
utilties)}

\item{p.noise}{The odds of randomly choosing a winner (instead of using
the utilities)}

\item{use.MNL}{Not used currently, placeholder for future.}

\item{verbose}{Whether to output information as it runs.}

\item{vec.format}{"WIN" will mark the winning row in the design matrix
with a "1" and all others with "0". This would match a typical regression
model to estimate coefficients from the observations.
"ANS" will put the row number of the winning concept into the rows for
each trial; some other software expects that format.}
}
\value{
A vector identifying which concept won each trial, according to the
design matrix and supplied utilities.
}
\description{
Choose the winning concepts in a CBC design matrix, according to utilities
}
\seealso{
[generateMNLrandomTab] to create a design matrix,
[generateRNDpws] to create a random set of utilities.
}
