% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdiff-plots.R
\name{plot.md.counts}
\alias{plot.md.counts}
\title{Plot best, worst, and net counts from MaxDiff data}
\usage{
\method{plot}{md.counts}(md.define, item.disguise = FALSE)
}
\arguments{
\item{md.define}{A MaxDiff study object with \code{$md.block} data,
as imported by \code{read.md.qualtrics()} or \code{read.md.cho()}.}

\item{item.disguise}{Remove the item labels and number them generically.
This is included to make it easy to share and show results at a conference
or other audience when the underlying details are confidential.}
}
\value{
A ggplot2 chart with the best, worst, and net counts.
}
\description{
This plot shows how often each item was chosen as best and as worst, and
shows the difference between those (the net count).
This is a simple descriptive alternative to logit model estimation
of MaxDiff values and generally corresponds extremely closely with
aggregate and upper-level model estimates from multinomial regression
and hierarchical Bayes estimation.
}
\details{
In the case that some items were shown more or less often than others,
regardless of whether they were chosen as best or worst or neither,
the counts here are rescaled to make the relative proportions equivalent.
For example, if one item appeared 1000 times but other items appeared only
800 times each, then the counts for those other items would be multiplied by
1.25 to account for the effective "proportion of times shown", compared
to the item with 1000 appearances.
}
