% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdiff-plots.R
\name{plot.md.indiv}
\alias{plot.md.indiv}
\title{Plot individual-level estimates (density) for MaxDiff utilites}
\usage{
\method{plot}{md.indiv}(
  md.define,
  use.raw = FALSE,
  item.disguise = FALSE,
  plot.zero = FALSE,
  plot.density = TRUE,
  plot.mean = TRUE,
  ran.seed = 98103
)
}
\arguments{
\item{md.define}{A MaxDiff study object that contains individual-level
utility estimates present in a member called \code{$md.hb.betas} or
\code{$md.hb.betas.zc}.
Typically those are estimated created by \code{md.hb()} in this package.

These data also be imported or created
by any other process, as long as they are rectangular with items in the
columns and respondents in the rows, and there is an "ID" column with
unique respondent identifies for each row. For example, if you export
data from Sawtooth Software or some other platform, you could load
and plot it with this function; just create an object
\code{md.define$md.hb.betas} with the estimates and an ID column.}

\item{use.raw}{Whether to use raw MNL beta coefficients \code{$md.hb.betas},
if present, or some transformed e.g., zero-centered diff, version of them
present in \code{$md.hb.betas}}

\item{item.disguise}{Remove the item labels and number them generically.
This is included to make it easy to share and show results at a conference
or other audience when the underlying details are confidential.}

\item{plot.zero}{Plot a vertical line for zero (the point of indifference),
which only applies for raw beta utilities.}

\item{plot.density}{Whether to draw density lines}

\item{plot.mean}{Whether to add the sample mean value as a red box}

\item{ran.seed}{RNG seed for the minor jittering that is added to the plot
for better filling of the density areas.}
}
\value{
A ggplot2 object with the plot.
}
\description{
This plot shows each individual's estimates as a blue circle for
each item, and allows visual inspection of the degree of dispersion
and general density properties of the estimates.
The sample level mean value is plotted as a red box.
}
