% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poetest.R
\name{poetest}
\alias{poetest}
\title{Perform the Poe (2005) test on different discrete choice models}
\usage{
poetest(n, model1, model2, att, price, vcov)
}
\arguments{
\item{n}{number of draws}

\item{model1}{Model from which to take the first WTP values}

\item{model2}{Model from which to take the second WTP values}

\item{att}{Vector of attributes whose WTP values you want to compare}

\item{price}{name of the price coefficient}

\item{vcov}{which variance-covariance matrix to use. Either normal for the normal one, or rob for the robust one.}
}
\value{
a p value associated with "WTP1>WTP2"
}
\description{
Perform the Poe (2005) test on different discrete choice models
}
\examples{
\dontrun{
poeresults<-poetest(n=5000, model1 = clmodels[[model_1]],model2 = clmodels[[model_2]],
 att=attr, price = "bcost", vcov = "normal")
}
}
