% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdiff-data-importexport.R
\name{preprocess.md.qualtrics}
\alias{preprocess.md.qualtrics}
\title{Preprocess Qualtrics observations and convert to a standard format.}
\usage{
preprocess.md.qualtrics(dat, itemConfirm = NULL)
}
\arguments{
\item{dat}{The data set read from Qualtrics by other functions.}

\item{itemConfirm}{A placeholder for a disambiguation string. Unused.}
}
\value{
Returns a data set in a consistent format for other functions in
\code{choicetools}.
}
\description{
This is an internal function. As Qualtrics changes data formats over time,
we convert them to a common format for sanity.
}
