% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicktexregapollo.R
\name{quicktexregapollo}
\alias{quicktexregapollo}
\title{Make your apollo object readable with texreg}
\usage{
quicktexregapollo(model = model, wtpest = NULL, se = "rob")
}
\arguments{
\item{model}{the name of the apollo model object}

\item{wtpest}{if you want to display wtp instead of beta coefficients provide a data.frame with the wtp values and standard errors}

\item{se}{Which standard errors should be used. Either rob for robust, normal for normal standard error or bs for bootstrapped standard errors.}
}
\value{
a list object to be easily used in texreg. It makes it easy to create tables usind different standard errors (including robust and bootstrapped) and to display WTP instead of beta coefficients.
}
\description{
Make your apollo object readable with texreg
}
\examples{
\dontrun{quicktexregapollo(model1)}
}
